rename to NestedTernaryToMatchRector
