+++ rules-tests/CodingStyle/Rector/Assign/NestedTernaryToMatchRector/Fixture/nested_ternary.php.inc
