+++ rules-tests/CodingStyle/Rector/Assign/NestedTernaryToMatchRector/Fixture/nested_ternary_with_identical.php.inc
