+++ rules-tests/CodingStyle/Rector/Assign/NestedTernaryToMatchRector/Fixture/skip_silent_conditions.php.inc
