+++ rules-tests/CodingStyle/Rector/Assign/NestedTernaryToMatchRector/config/configured_rule.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+use Rector\CodingStyle\Rector\Assign\NestedTernaryToMatchRector;
+use Rector\Config\RectorConfig;
+
+return RectorConfig::configure()
+    ->withRules([NestedTernaryToMatchRector::class]);
