--- rules-tests/CodingStyle/Rector/Assign/NestedTernaryToMatchTrueRector/config/configured_rule.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-use Rector\CodingStyle\Rector\Assign\NestedTernaryToMatchTrueRector;
-use Rector\Config\RectorConfig;
-
-return RectorConfig::configure()
-    ->withRules([NestedTernaryToMatchTrueRector::class]);
